// frontend/src/shared/api/generated/models/LotteryState.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Swagger TON Lottery
 * TON Lottery
 *
 * The version of the OpenAPI document: 1.0.0
 * Contact: almostinf@mail.ru
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * 
 * @export
 */
export const LotteryState = {
    LotteryStatePrepared: 'LOTTERY_STATE_PREPARED',
    LotteryStateActive: 'LOTTERY_STATE_ACTIVE',
    LotteryStateFinished: 'LOTTERY_STATE_FINISHED',
    LotteryStateFailed: 'LOTTERY_STATE_FAILED'
} as const;
export type LotteryState = typeof LotteryState[keyof typeof LotteryState];


export function instanceOfLotteryState(value: any): boolean {
    for (const key in LotteryState) {
        if (Object.prototype.hasOwnProperty.call(LotteryState, key)) {
            if (LotteryState[key as keyof typeof LotteryState] === value) {
                return true;
            }
        }
    }
    return false;
}

export function LotteryStateFromJSON(json: any): LotteryState {
    return LotteryStateFromJSONTyped(json, false);
}

export function LotteryStateFromJSONTyped(json: any, ignoreDiscriminator: boolean): LotteryState {
    return json as LotteryState;
}

export function LotteryStateToJSON(value?: LotteryState | null): any {
    return value as any;
}

export function LotteryStateToJSONTyped(value: any, ignoreDiscriminator: boolean): LotteryState {
    return value as LotteryState;
}