// frontend/src/shared/api/generated/models/TicketState.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Swagger TON Lottery
 * TON Lottery
 *
 * The version of the OpenAPI document: 1.0.0
 * Contact: almostinf@mail.ru
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * 
 * @export
 */
export const TicketState = {
    TicketStatePrepared: 'TICKET_STATE_PREPARED',
    TicketStateActive: 'TICKET_STATE_ACTIVE',
    TicketStateRefunded: 'TICKET_STATE_REFUNDED',
    TicketStateFailed: 'TICKET_STATE_FAILED',
    TicketStateClosed: 'TICKET_STATE_CLOSED'
} as const;
export type TicketState = typeof TicketState[keyof typeof TicketState];


export function instanceOfTicketState(value: any): boolean {
    for (const key in TicketState) {
        if (Object.prototype.hasOwnProperty.call(TicketState, key)) {
            if (TicketState[key as keyof typeof TicketState] === value) {
                return true;
            }
        }
    }
    return false;
}

export function TicketStateFromJSON(json: any): TicketState {
    return TicketStateFromJSONTyped(json, false);
}

export function TicketStateFromJSONTyped(json: any, ignoreDiscriminator: boolean): TicketState {
    return json as TicketState;
}

export function TicketStateToJSON(value?: TicketState | null): any {
    return value as any;
}

export function TicketStateToJSONTyped(value: any, ignoreDiscriminator: boolean): TicketState {
    return value as TicketState;
}